+++ tests/testthat/test_pipeop_imputelearner.R
@@ -172,3 +172,68 @@ test_that("marshal", {
   expect_equal(s, su)
 })
 
+test_that("Imputing zero level factors", {
+  # Tasks with zero levels cannot be directly created. However, they can occur within pipelines using FixFactors.
+  base_task = TaskClassif$new("test", target = "target", backend = data.table(
+    target = factor(c("C1", "C1", "C2", "C2", "C1")),
+    fct = factor(rep(NA, 5), levels = "lvl"),
+    ord = ordered(rep(NA, 5), levels = "lvl")
+  ))
+  task = po("fixfactors")$train(list(base_task))[[1L]]
+  # Just to be robust for possible changes to FixFactors in the future, since the usage here is a bit hacky
+  expect_equal(task$levels(), list(fct = character(0), ord = character(0), target = c("C1", "C2")))
+
+  op = po("imputelearner", learner = lrn("classif.featureless"))
+  expect_no_error({
+    expect_equal(op$train(list(task))[[1L]]$data(), task$data())
+  })
+  expect_no_error({
+    expect_equal(op$predict(list(task))[[1L]]$data(), task$data())
+  })
+
+})
+
+test_that("PipeOpImputeLearner - correct levels, #691", {
+  op = po("imputelearner", learner = lrn("classif.featureless"))
+  task = TaskRegr$new("test", target = "y", backend = data.table(
+    y = seq(1:5),
+    x1 = factor(c(NA, "a", "a", "b", "b")),
+    x2 = ordered(c(NA, "a", "a", "b", "b"))
+  ))
+
+  expect_equal(
+    op$train(list(task))[[1L]]$levels(),
+    list(x1 = c("a", "b"), x2 = c("a", "b"))
+  )
+  expect_equal(
+    op$predict(list(task))[[1L]]$levels(),
+    list(x1 = c("a", "b"), x2 = c("a", "b"))
+  )
+
+})
+
+test_that("PipeOpImputeLearner - impute missings for unseen factor levels", {
+  skip_if_not_installed("rpart")
+  # Construct Learner incapable of handling missings
+  learner = lrn("classif.rpart")
+  learner$properties = setdiff(learner$properties, "missings")
+  # Construct Tasks with unseen factor levels
+  task_NA = as_task_classif(data.table(
+    target = factor(rep(c("A", "B"), 3)),
+    fct = factor(rep(c("a", "b", NA), 2))
+  ), target = "target")
+  task_noNA = as_task_classif(data.table(
+    target = factor(rep(c("A", "B"), 3)),
+    fct = factor(rep(c("a", "b", "c"), 2))
+  ), target = "target")
+
+  op = po("imputelearner", learner = lrn("classif.featureless"))
+
+  expect_equal(sum(op$train(list(task_NA))[[1]]$missings()), 0)
+  expect_equal(sum(op$predict(list(task_noNA))[[1]]$missings()), 0)
+
+  glrn = op %>>% learner
+  expect_no_error(glrn$train(task_NA))
+  expect_no_error(glrn$predict(task_NA))
+
+})