+++ tests/testthat/test_pipeop_textvectorizer.R
@@ -2,7 +2,6 @@ context("PipeOpTextVectorizer")
 
 test_that("PipeOpTextVectorizer - basic properties", {
   skip_if_not_installed("quanteda")
-  suppressWarnings(loadNamespace("quanteda"))  # TODO: see https://github.com/quanteda/quanteda/issues/2116 , may not be an issue in the future
 
   task = mlr_tasks$get("iris")
   # create hacky text data:
@@ -45,7 +44,6 @@ test_that("PipeOpTextVectorizer - basic properties", {
 
 test_that("PipeOpTextVectorizer - tfidf works", {
   skip_if_not_installed("quanteda")
-  suppressWarnings(loadNamespace("quanteda"))  # TODO: see https://github.com/quanteda/quanteda/issues/2116 , may not be an issue in the future
 
   task = mlr_tasks$get("iris")
   # create some text data
@@ -116,7 +114,6 @@ test_that("PipeOpTextVectorizer - tfidf works", {
 
 test_that("PipeOpTextVectorizer - bigrams", {
   skip_if_not_installed("quanteda")
-  suppressWarnings(loadNamespace("quanteda"))  # TODO: see https://github.com/quanteda/quanteda/issues/2116 , may not be an issue in the future
 
   task = mlr_tasks$get("iris")
   # create hacky text data:
@@ -153,7 +150,6 @@ test_that("PipeOpTextVectorizer - bigrams", {
 
 test_that("PipeOpTextVectorizer - integer sequence", {
   skip_if_not_installed("quanteda")
-  suppressWarnings(loadNamespace("quanteda"))  # TODO: see https://github.com/quanteda/quanteda/issues/2116 , may not be an issue in the future
 
   set.seed(1)
   task = mlr_tasks$get("iris")
@@ -225,7 +221,6 @@ test_that("PipeOpTextVectorizer - integer sequence", {
 
 test_that("PipeOpTextVectorizer - factor sequence", {
   skip_if_not_installed("quanteda")
-  suppressWarnings(loadNamespace("quanteda"))  # TODO: see https://github.com/quanteda/quanteda/issues/2116 , may not be an issue in the future
 
   task = mlr_tasks$get("iris")
   # create hacky text data: